requirements.txt
openpyxl==3.1.2
pandas==1.5.3
psycopg2_binary==2.9.9
streamlit==1.23.1
streamlit_timeline==0.0.2
